move templating types into one file
