+++ packages/insomnia/src/common/har.ts
@@ -12,13 +12,14 @@ import { getAuthHeader } from '../network/authentication';
 import * as plugins from '../plugins';
 import * as pluginContexts from '../plugins/context/index';
 import { RenderError } from '../templating/render-error';
+import type { RenderedRequest } from '../templating/types';
 import { parseGraphQLReqeustBody } from '../utils/graph-ql';
 import { smartEncodeUrl } from '../utils/url/querystring';
 import { getAppVersion } from './constants';
 import { jarFromCookies } from './cookies';
 import { database } from './database';
 import { filterHeaders, getSetCookieHeaders, hasAuthHeader } from './misc';
-import { getRenderedRequestAndContext, type RenderedRequest } from './render';
+import { getRenderedRequestAndContext } from './render';
 
 export interface ExportRequest {
   requestId: string;
