+++ packages/insomnia/src/plugins/context/network.ts
@@ -1,7 +1,7 @@
-import { type ExtraRenderInfo } from '../../common/render';
 import * as models from '../../models';
 import type { Request } from '../../models/request';
 import { fetchRequestData, responseTransform, sendCurlAndWriteTimeline, tryToInterpolateRequest, tryToTransformRequestWithPlugins } from '../../network/network';
+import { type ExtraRenderInfo } from '../../templating/types';
 
 export function init() {
   return {
