+++ packages/insomnia/src/plugins/context/request.ts
@@ -1,6 +1,6 @@
 import * as misc from '../../common/misc';
-import type { RenderedRequest } from '../../common/render';
 import type { RequestBody } from '../../models/request';
+import type { RenderedRequest } from '../../templating/types';
 export function filterParameters<T extends { name: string; value: string }>(
   parameters: T[],
   name: string,
