+++ packages/insomnia/src/ui/components/markdown-preview.tsx
@@ -2,7 +2,7 @@ import highlight from 'highlight.js/lib/common';
 import React, { type FC, useEffect, useLayoutEffect, useRef, useState } from 'react';
 
 import { markdownToHTML } from '../../common/markdown-to-html';
-import type { HandleRender } from '../../common/render';
+import type { HandleRender } from '../../templating/types';
 
 interface Props {
   markdown: string;
