+++ packages/insomnia/src/utils/graph-ql.ts
@@ -1,8 +1,8 @@
 import { getOperationAST, parse } from 'graphql';
 
 import { CONTENT_TYPE_GRAPHQL } from '../common/constants';
-import type { RenderedRequest } from '../common/render';
 import type { Request } from '../models/request';
+import type { RenderedRequest } from '../templating/types';
 
 // parse graphql request body since we save entire query variables as string rather then stringified json string. - INS-4281
 export function parseGraphQLReqeustBody(renderedRequest: RenderedRequest) {
